+++ physical_ai_server/physical_ai_server/communication/communicator.py
@@ -19,6 +19,8 @@
 from functools import partial
 from typing import Any, Dict, Optional, Set, Tuple
 
+from geometry_msgs.msg import Twist
+from nav_msgs.msg import Odometry
 from physical_ai_interfaces.msg import TaskStatus
 from physical_ai_interfaces.srv import (
     GetImageTopicList
@@ -40,6 +42,8 @@ class Communicator:
     # Define operation modes
     MODE_COLLECTION = 'collection'  # Full data collection mode (images, follower, leader)
     MODE_INFERENCE = 'inference'    # Inference mode (images, follower only)
+
+    PUB_QOS_SIZE = 100
 
     def __init__(
         self,
@@ -106,12 +110,18 @@ class Communicator:
         for name, topic in self.joint_topics.items():
             # Determine category and message type based on name patterns
             if 'follower' in name.lower():
+                if 'mobile' in name.lower():
+                    msg_type = Odometry
+                else:
+                    msg_type = JointState
                 category = self.SOURCE_FOLLOWER
-                msg_type = JointState
                 callback = partial(self._follower_callback, name)
             elif 'leader' in name.lower():
+                if 'mobile' in name.lower():
+                    msg_type = Twist
+                else:
+                    msg_type = JointTrajectory
                 category = self.SOURCE_LEADER
-                msg_type = JointTrajectory
                 callback = partial(self._leader_callback, name)
             else:
                 # Log an error message if the topic name does not include 'follower' or 'leader'
@@ -127,25 +137,31 @@ class Communicator:
                 msg_type=msg_type,
                 callback=callback
             )
-            self.joint_topics[name] = msg_type()
             self.node.get_logger().info(
                 f'Joint subscriber: {name} -> {topic} ({msg_type.__name__})')
 
     def init_publishers(self):
-        # TODO: Re-enable the code below in a future PR
-        # to implement joint control.
-        # for name, topic_name in self.joint_topics.items():
-        #     if 'leader' in name.lower():
-        #         self.joint_publishers[name] = self.node.create_publisher(
-        #             JointTrajectory,
-        #             topic_name,
-        #             100
-        #         )
+        self.node.get_logger().info('Initializing joint publishers...')
+        for name, topic_name in self.joint_topics.items():
+            if 'leader' in name.lower():
+                if 'mobile' in name.lower():
+                    self.joint_publishers[name] = self.node.create_publisher(
+                        Twist,
+                        topic_name,
+                        self.PUB_QOS_SIZE
+                    )
+                else:
+                    self.joint_publishers[name] = self.node.create_publisher(
+                        JointTrajectory,
+                        topic_name,
+                        self.PUB_QOS_SIZE
+                    )
+        self.node.get_logger().info('Initializing joint publishers... done')
 
         self.status_publisher = self.node.create_publisher(
             TaskStatus,
             '/task/status',
-            100
+            self.PUB_QOS_SIZE
         )
 
     def init_services(self):
@@ -172,8 +188,8 @@ class Communicator:
         elif self.operation_mode == self.MODE_INFERENCE:
             return self.camera_topic_msgs, self.follower_topic_msgs, None
 
-    def send_action(self, joint_msgs: Dict[str, JointTrajectory]):
-        for name, joint_msg in joint_msgs.items():
+    def publish_action(self, joint_msg_datas: Dict[str, Any]):
+        for name, joint_msg in joint_msg_datas.items():
             self.joint_publishers[name].publish(joint_msg)
 
     def publish_status(self, status: TaskStatus):
@@ -188,7 +204,7 @@ class Communicator:
             camera_topic_list.append(topic)
 
         if len(camera_topic_list) == 0:
-            self.get_logger().error('No image topics found')
+            self.node.get_logger().error('No image topics found')
             response.image_topic_list = []
             response.success = False
             response.message = 'Please check image topics in your robot configuration.'
@@ -198,6 +214,12 @@ class Communicator:
         response.success = True
         response.message = 'Image topic list retrieved successfully'
         return response
+
+    def get_publisher_msg_types(self):
+        msg_types = {}
+        for publisher_name, publisher in self.joint_publishers.items():
+            msg_types[publisher_name] = publisher.msg_type
+        return msg_types
 
     def cleanup(self):
         self.node.get_logger().info(